Moved parts of the program class to ReleaseNotesGenerator in preparation to usage outside console apps
